backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
elasticsearch>=8.0.0,<9.0.0
python-dotenv
python-multipart
jinja2
